feat: lsb1 embed implementation

+++ src/algorithms/stego_algorithms.h
@@ -1,7 +1,10 @@
 #ifndef STEGO_ALGORITHMS_H
 #define STEGO_ALGORITHMS_H
 
-void lsb1_embed(uint8_t * carrier, size_t * carrier_size, uint8_t * hide, size_t * hide_size);
+#include <stdint.h>
+
+
+void lsb1_embed(uint8_t * carrier, uint32_t * carrier_size, uint8_t * hide, uint32_t * hide_size);
 
 void lsb1_extract();
